+++ smol-core/src/Smol/Core/Typecheck/Typeclass/ToDictionaryPassing.hs
@@ -10,6 +10,7 @@ module Smol.Core.Typecheck.Typeclass.ToDictionaryPassing
     toDictionaryPassing,
     passDictionaries,
     lookupTypecheckedTypeclassInstance,
+    ToDictEnv (..),
   )
 where
 
@@ -34,20 +35,16 @@ import Smol.Core.Types
 lookupTypecheckedTypeclassInstance ::
   forall m ann.
   (MonadError (TCError ann) m, Monoid ann, Ord ann, Show ann) =>
-  M.Map TypeclassName (Typeclass ResolvedDep ann) ->
-  M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)) ->
+  ToDictEnv ann ->
   Constraint ResolvedDep (Type ResolvedDep ann) ->
   m ([Substitution ResolvedDep ann], Instance ResolvedDep (Type ResolvedDep ann))
-lookupTypecheckedTypeclassInstance typeClasses instances constraint = do
-  tracePrettyM "lookupTypecheckedTypeclassInstance" constraint
+lookupTypecheckedTypeclassInstance env constraint = do
+  let instances = tdeInstances env
   case M.lookup (void constraint) instances of
-    Just tcInstance -> do
-      tracePrettyM "found concrete instance" tcInstance
-      pure (mempty, tcInstance)
+    Just tcInstance -> pure (mempty, tcInstance)
     Nothing -> do
       (foundConstraint, subs) <-
         findMatchingConstraint (M.keys instances) (removeTypesFromConstraint constraint)
-      tracePrettyM "foundConstraint, subs" (foundConstraint, subs)
 
       -- we can't actually provide an instance so what to do?
       -- a) look up main instance
@@ -59,10 +56,8 @@ lookupTypecheckedTypeclassInstance typeClasses instances constraint = do
                   . removeTypesFromConstraint
                   <$> inConstraints
 
-          tracePrettyM "subbed instance" (subbedConstraints, inExpr)
-
           -- see if found types exist
-          traverse_ (lookupTypecheckedTypeclassInstance typeClasses instances) (addTypesToConstraint <$> subbedConstraints)
+          traverse_ (lookupTypecheckedTypeclassInstance env) (addTypesToConstraint <$> subbedConstraints)
 
           -- return new instance
           pure
@@ -110,19 +105,18 @@ getTypeForDictionary ::
     Ord ann,
     Show ann
   ) =>
-  M.Map TypeclassName (Typeclass ResolvedDep ann) ->
-  M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)) ->
+  ToDictEnv ann ->
   [Constraint ResolvedDep (Type ResolvedDep ann)] ->
   m (Maybe (Pattern ResolvedDep (Type ResolvedDep ann)))
-getTypeForDictionary typeClasses instances constraints = do
+getTypeForDictionary env constraints = do
   let getConstraintPattern constraint i = do
         let ident = identForConstraint (i + 1)
-        ty <- case lookupTypecheckedTypeclassInstance typeClasses instances constraint of
+        ty <- case lookupTypecheckedTypeclassInstance env constraint of
           -- we found the instance, return it's type
           Right (_, Instance _ instanceExpr) -> pure (getExprAnnotation instanceExpr)
           -- we didn't find an instance, but we can get the type from the
           -- constraint
-          Left e -> case typeForConstraint typeClasses constraint of
+          Left e -> case typeForConstraint (tdeClasses env) constraint of
             Just ty -> pure ty
             Nothing -> throwError e
         pure (PVar ty ident)
@@ -150,17 +144,14 @@ typeForConstraint typeClasses constraint@(Constraint tcn _) = do
 -- `let equals_1 = \a -> \b -> a == b in equals_1 10 11`
 convertExprToUseTypeclassDictionary ::
   (MonadError (TCError ann) m, Monoid ann, Ord ann, Show ann) =>
-  M.Map TypeclassName (Typeclass ResolvedDep ann) ->
-  M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)) ->
+  ToDictEnv ann ->
   [Constraint ResolvedDep (Type ResolvedDep ann)] ->
   Expr ResolvedDep (Type ResolvedDep ann) ->
   m (Expr ResolvedDep (Type ResolvedDep ann))
-convertExprToUseTypeclassDictionary typeClasses instances constraints expr = do
-  tracePrettyM "convertExprToUseTypeclassDictionary" (constraints, expr)
-
+convertExprToUseTypeclassDictionary env constraints expr = do
   -- if our constraints are concrete we'll inline them rather than passing them
   -- through, as such
-  maybePattern <- getTypeForDictionary typeClasses instances (filterNotConcrete constraints)
+  maybePattern <- getTypeForDictionary env (filterNotConcrete constraints)
 
   case maybePattern of
     Just pat -> do
@@ -182,24 +173,23 @@ convertExprToUseTypeclassDictionary typeClasses instances constraints expr = do
 -- (ie "pass them through", as such)
 createTypeclassDict ::
   (Show ann, Ord ann, Monoid ann, MonadError (TCError ann) m) =>
-  TCEnv ann ->
-  M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)) ->
+  ToDictEnv ann ->
   NE.NonEmpty (Constraint ResolvedDep (Type ResolvedDep ann)) ->
   m (Expr ResolvedDep (Type ResolvedDep ann))
-createTypeclassDict env instances constraints = do
+createTypeclassDict env constraints = do
   foundInstances <-
     traverse
       ( \constraint -> do
-          case lookupTypecheckedTypeclassInstance (tceClasses env) instances constraint of
+          case lookupTypecheckedTypeclassInstance env constraint of
             Right (subs, Instance newConstraints expr) -> do
               -- found a concrete instance
-              toDictionaryPassing (tceClasses env) instances subs newConstraints expr
+              toDictionaryPassing env subs newConstraints expr
             Left e -> do
               -- no concrete instance, maybe we can pass through a constraint
               -- from the current function
               case (,)
-                <$> elemIndex (removeTypesFromConstraint constraint) (tceConstraints env)
-                <*> typeForConstraint (tceClasses env) constraint of
+                <$> elemIndex (removeTypesFromConstraint constraint) (removeTypesFromConstraint <$> NE.toList constraints)
+                <*> typeForConstraint (tdeClasses env) constraint of
                 Just (index, ty) -> pure (EVar ty (identForConstraint $ fromIntegral index))
                 Nothing -> throwError e
       )
@@ -217,49 +207,56 @@ filterNotConcrete = filter (not . isConcrete)
 -- pass dictionaries to them all
 passDictionaries ::
   (Monoid ann, Ord ann, Show ann, MonadError (TCError ann) m) =>
-  M.Map TypeclassName (Typeclass ResolvedDep ann) ->
-  M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)) ->
+  ToDictEnv ann ->
   [Substitution ResolvedDep ann] ->
   Expr ResolvedDep (Type ResolvedDep ann) ->
   m (Expr ResolvedDep (Type ResolvedDep ann))
-passDictionaries typeClasses instances subs =
+passDictionaries env subs =
   go
   where
-    go (EVar ann ident) = do
-      tracePrettyM "passDictionary to " ident
-      result <- recoverInstance typeClasses ident ann
-      case result of
-        Just constraint -> do
-          tracePrettyM "recovered an instance" (ident, constraint)
-
-          -- specialise contraints to found types
-          let subbedConstraint =
-                substituteConstraint subs constraint
-
-          (newSubs, Instance fnConstraints fnExpr) <-
-            liftEither (lookupTypecheckedTypeclassInstance typeClasses instances (addTypesToConstraint subbedConstraint))
-
-          tracePrettyM "found instance" (fnConstraints, fnExpr)
-
-          -- convert instance to dictionary passing then return it inlined
-          toDictionaryPassing typeClasses instances newSubs fnConstraints fnExpr
+    go (EVar ann ident) = 
+      case M.lookup ident (tdeVars env) of
+        Just (constraints, _defExpr) -> do
+          -- need to specialise constraint to actual type here
+          case NE.nonEmpty constraints of
+            Just neConstraints -> do
+              -- use the call type to specialise to the instance we need
+              specialisedConstraints <- traverse (specialiseConstraint (tdeClasses env) ann) neConstraints
+              EApp ann (EVar ann ident) <$> createTypeclassDict env (addTypesToConstraint <$> specialisedConstraints)
+            Nothing -> pure (EVar ann ident)
         Nothing -> do
-          tracePrettyM "nothng recovered for " ident
-          pure (EVar ann ident)
+          result <- recoverInstance (tdeClasses env) ident ann
+          case result of
+            Just constraint -> do
+
+              -- specialise contraints to found types
+              let subbedConstraint =
+                    substituteConstraint subs constraint
+
+              (newSubs, Instance fnConstraints fnExpr) <-
+                liftEither (lookupTypecheckedTypeclassInstance env (addTypesToConstraint subbedConstraint))
+
+              -- convert instance to dictionary passing then return it inlined
+              toDictionaryPassing env newSubs fnConstraints fnExpr
+            Nothing -> pure (EVar ann ident)
     go other = bindExpr go other
+
+data ToDictEnv ann = ToDictEnv
+  { tdeClasses :: M.Map TypeclassName (Typeclass ResolvedDep ann),
+    tdeInstances :: M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)),
+    tdeVars :: M.Map (ResolvedDep Identifier) ([Constraint ResolvedDep ann], ResolvedType ann)
+  }
 
 -- | well well well lets put it all together
 toDictionaryPassing ::
   (MonadError (TCError ann) m, Show ann, Ord ann, Monoid ann) =>
-  M.Map TypeclassName (Typeclass ResolvedDep ann) ->
-  M.Map (Constraint ResolvedDep ()) (Instance ResolvedDep (Type ResolvedDep ann)) ->
+  ToDictEnv ann ->
   [Substitution ResolvedDep ann] ->
   [Constraint ResolvedDep (Type ResolvedDep ann)] ->
   Expr ResolvedDep (Type ResolvedDep ann) ->
   m (Expr ResolvedDep (Type ResolvedDep ann))
-toDictionaryPassing typeClasses instances subs constraints expr = do
-  tracePrettyM "toDictionaryPassing" (subs, constraints, expr)
+toDictionaryPassing env subs constraints expr = do
 
-  passDictionaries typeClasses instances subs
-    <=< convertExprToUseTypeclassDictionary typeClasses instances constraints
+  passDictionaries env subs
+    <=< convertExprToUseTypeclassDictionary env constraints
     $ expr
