+++ smol-core/test/Test/Typecheck/TypeclassSpec.hs
@@ -7,81 +7,12 @@
 
 module Test.Typecheck.TypeclassSpec (spec) where
 
-import Data.Bifunctor (bimap)
 import Data.Either
-import Data.Foldable (traverse_)
-import Data.Functor
-import Data.List (nub)
-import qualified Data.List.NonEmpty as NE
 import qualified Data.Map.Strict as M
-import Data.String (fromString)
-import qualified Data.Text as T
 import Smol.Core
-import Smol.Core.Modules.ResolveDeps
-import Smol.Core.Modules.Types.Module
-import Smol.Core.Typecheck.FromParsedExpr (fromParsedExpr)
 import Smol.Core.Typecheck.Typeclass
 import Test.Helpers
 import Test.Hspec
-
-simplify :: Expr ResolvedDep ann -> Expr ResolvedDep ()
-simplify = void . goExpr
-  where
-    changeIdent (TypeclassCall ident i) =
-      LocalDefinition $ "tc" <> ident <> fromString (show i)
-    changeIdent (UniqueDefinition ident i) =
-      LocalDefinition $ ident <> fromString (show i)
-    changeIdent ident = ident
-
-    goExpr (EVar ann ident) =
-      EVar ann (changeIdent ident)
-    goExpr (EAnn ann ty rest) =
-      EAnn ann (typeForComparison ty) (goExpr rest)
-    goExpr (ELambda ann ident body) =
-      ELambda ann (changeIdent ident) (goExpr body)
-    goExpr (EPatternMatch ann matchExpr pats) =
-      EPatternMatch ann (goExpr matchExpr) (fmap (bimap goPattern goExpr) pats)
-    goExpr other = mapExpr goExpr other
-
-    goPattern (PVar ann ident) = PVar ann (changeIdent ident)
-    goPattern other = mapPattern goPattern other
-
-evalExpr ::
-  [Constraint ResolvedDep Annotation] ->
-  T.Text ->
-  Either
-    (TCError Annotation)
-    ( ResolvedExpr (Type ResolvedDep Annotation),
-      M.Map (ResolvedDep Identifier) (Constraint ResolvedDep Annotation)
-    )
-evalExpr constraints input =
-  case parseExprAndFormatError input of
-    Left e -> error (show e)
-    Right expr ->
-      case resolveExprDeps expr (getTypeclassMethodNames @() typecheckEnv) mempty of
-        Left e -> error $ "error resolving Expr deps :" <> show e
-        Right resolvedExpr ->
-          let env =
-                typecheckEnv
-                  { tceConstraints = constraints
-                  }
-           in elaborate env resolvedExpr
-
--- | elaborate but don't do clever resolving so we can construct the
--- expectations we want
-evalExprUnsafe ::
-  T.Text ->
-  Either (TCError Annotation) (ResolvedExpr (Type ResolvedDep Annotation))
-evalExprUnsafe input = case parseExprAndFormatError input of
-  Left e -> error (show e)
-  Right expr ->
-    case elaborate typecheckEnv (fromParsedExpr expr) of
-      Right (typedExpr, _typeclassUses) -> pure typedExpr
-      Left e -> Left e
-
-getRight :: (Show e) => Either e a -> a
-getRight (Right a) = a
-getRight (Left e) = error (show e)
 
 spec :: Spec
 spec = do
@@ -253,102 +184,9 @@ spec = do
           tyTypeclass = unsafeParseType "Maybe a"
       matchType tyMatch tyTypeclass `shouldSatisfy` isRight
 
-  describe "Get dictionaries" $ do
-    it "Single item dictionary for single constraint" $ do
-      let constraints = addTypesToConstraint <$> NE.fromList [Constraint "Eq" [tyInt]]
-          expected = evalExprUnsafe "(\\a1 -> \\b2 -> a1 == b2 : Int -> Int -> Bool)"
-
-          instances = moInstances testModule
-
-      fmap simplify (createTypeclassDict typecheckEnv instances constraints)
-        `shouldBe` simplify <$> expected
-
-    it "Tuple for two constraints" $ do
-      let constraints = addTypesToConstraint <$> NE.fromList [Constraint "Eq" [tyInt], Constraint "Eq" [tyInt]]
-          expected = evalExprUnsafe "((\\a1 -> \\b2 -> a1 == b2 : Int -> Int -> Bool), (\\a1 -> \\b2 -> a1 == b2 : Int -> Int -> Bool))"
-
-          instances = moInstances testModule
-
-      fmap simplify (createTypeclassDict typecheckEnv instances constraints)
-        `shouldBe` simplify <$> expected
-
   describe "isConcrete" $ do
     it "yes, because it has no vars" $ do
       isConcrete @_ @() (Constraint "Eq" [tyInt]) `shouldBe` True
 
     it "no, because it has a var" $ do
       isConcrete @_ @() (Constraint "Eq" [tcVar "a"]) `shouldBe` False
-
-  describe "Convert expr to use typeclass dictionaries" $ do
-    traverse_
-      ( \(constraints, parts, expectedConstraints, expectedParts) ->
-          let input = joinText parts
-              expected = joinText expectedParts
-           in it ("Successfully converted " <> show input) $ do
-                let (expr, typeclassUses) = getRight $ evalExpr constraints input
-                    instances = mempty
-
-                let expectedExpr = getRight $ evalExprUnsafe expected
-                    (dedupedConstraints, tidyExpr) = deduplicateConstraints typeclassUses expr
-                    result = convertExprToUseTypeclassDictionary (tceClasses typecheckEnv) instances (addTypesToConstraint <$> dedupedConstraints) tidyExpr
-
-                dedupedConstraints `shouldBe` expectedConstraints
-                simplify <$> result `shouldBe` Right (simplify expectedExpr)
-      )
-      [ (mempty, ["1 + 2"], mempty, ["1 + 2"]),
-        ( [ Constraint "Eq" [tcVar "a"],
-            Constraint "Eq" [tcVar "b"]
-          ],
-          [ "(\\a -> \\b -> case (a,b) of ((leftA, leftB), (rightA, rightB)) -> ",
-            "if equals leftA rightA then equals leftB rightB else False : (a,b) -> (a,b) -> Bool)"
-          ],
-          [ Constraint "Eq" [tcVar "a"],
-            Constraint "Eq" [tcVar "b"]
-          ],
-          [ "\\instances -> case (instances : (a -> a -> Bool, b -> b -> Bool)) of (tcvaluefromdictionary0, tcvaluefromdictionary1) -> ",
-            "(\\a1 -> \\b2 -> case (a1,b2) of ((leftA3, leftB4), (rightA5, rightB6)) ->",
-            "if tcvaluefromdictionary0 leftA3 rightA5 then tcvaluefromdictionary1 leftB4 rightB6 else False : (a,b) -> (a,b) -> Bool)"
-          ]
-        )
-      ]
-
-  -- the whole transformation basically
-  describe "toDictionaryPassing" $ do
-    traverse_
-      ( \(constraints, parts, expectedParts) -> do
-          let input = joinText parts
-              expected = joinText expectedParts
-           in it ("Successfully inlined " <> show input) $ do
-                let (expr, typeclassUses) = getRight $ evalExpr constraints input
-
-                    instances = moInstances testModule
-
-                let expectedExpr = getRight $ evalExprUnsafe expected
-                    (dedupedConstraints, tidyExpr) = deduplicateConstraints typeclassUses expr
-                    allConstraints = nub (dedupedConstraints <> constraints) -- we lose outer constraints sometimes
-                    result = toDictionaryPassing (tceClasses typecheckEnv) instances mempty (addTypesToConstraint <$> allConstraints) tidyExpr
-
-                simplify <$> result `shouldBe` Right (simplify expectedExpr)
-      )
-      [ (mempty, ["1 + 2"], ["1 + 2"]),
-        ( mempty,
-          ["equals (1: Int) (2: Int)"],
-          [ "(\\a1 -> \\b2 -> a1 == b2 : Int -> Int -> Bool) (1 : Int) (2: Int)"
-          ]
-        ),
-        ( mempty,
-          ["equals ((1: Int), (2: Int)) ((2: Int), (3: Int))"],
-          [ "(\\pairA7 -> \\pairB8 -> case (pairA7, pairB8) of ((a19, b110), (a211, b212)) ->",
-            "if (\\a1 -> \\b2 -> a1 == b2 : Int -> Int -> Bool) a19 a211 ",
-            "then (\\a1 -> \\b2 -> a1 == b2 : Int -> Int -> Bool) b110 b212",
-            "else False : (a, b) -> (a,b) -> Bool)",
-            "((1: Int), (2: Int)) ((2: Int), (3: Int))"
-          ]
-        ),
-        ( [Constraint "Eq" [tcVar "a"]],
-          ["(\\a -> \\b -> equals a b : a -> a -> Bool)"],
-          [ "\\instances -> case (instances : (a -> a -> Bool)) of tcvaluefromdictionary0 -> ",
-            "(\\a1 -> \\b2 -> tcvaluefromdictionary0 a1 b2 : a -> a -> Bool)"
-          ]
-        )
-      ]
